requirements.txt
pip-tools
cookiecutter
flake8
pytest
pytest-black
pytest-django
pytest-isort
pytest-flake8
pytest-mypy
sh
tox
virtualenv